SL-1348 SL-1349: changed multi sessionTypes to one sessiontype for Session table, fixed unit tests, removed unused setter
