+++ src/main/java/uk/gov/hmcts/reform/sandl/snlevents/model/db/Session.java
@@ -1,6 +1,7 @@
 package uk.gov.hmcts.reform.sandl.snlevents.model.db;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
+import com.fasterxml.jackson.annotation.JsonInclude;
 import lombok.AllArgsConstructor;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
@@ -10,14 +11,11 @@ import lombok.Setter;
 import java.io.Serializable;
 import java.time.Duration;
 import java.time.OffsetDateTime;
-import java.util.HashSet;
 import java.util.List;
-import java.util.Set;
 import java.util.UUID;
 import javax.persistence.CascadeType;
 import javax.persistence.Entity;
 import javax.persistence.Id;
-import javax.persistence.ManyToMany;
 import javax.persistence.ManyToOne;
 import javax.persistence.OneToMany;
 import javax.validation.constraints.NotNull;
@@ -63,14 +61,15 @@ public class Session extends VersionedEntity implements Serializable {
     private List<HearingPart> hearingParts;
 
     @Getter
-    @ManyToMany(cascade = {
+    @JsonInclude(JsonInclude.Include.NON_NULL)
+    @ManyToOne(cascade = {
         CascadeType.PERSIST,
         CascadeType.MERGE
-        }, mappedBy = "sessions")
-    private Set<SessionType> sessionTypes = new HashSet<>();
+        })
+    private SessionType sessionType;
 
-    public void addSessionType(SessionType sessionType) {
-        sessionTypes.add(sessionType);
+    public void setSessionType(SessionType sessionType) {
+        this.sessionType = sessionType;
         sessionType.getSessions().add(this);
     }
 }
